refacto(admin): extract ProductSection and UserSection components from AdminView #30
