+++ frontend/src/components/admin/AddProductForm.vue
