+++ frontend/src/components/admin/UserSection.vue
@@ -1,0 +1,223 @@
+<template>
+  <div>
+    <div class="content-header">
+      <h1 class="content-title">Gestion utilisateurs</h1>
+      <button class="btn-add" @click="showAddUserModal = true">
+        + Ajouter un utilisateur
+      </button>
+    </div>
+
+    <MessageBox v-if="successMessage" type="info" :message="successMessage" />
+    <MessageBox v-if="errorMessage" type="error" :message="errorMessage" />
+
+    <div v-if="loading" class="state-message">Chargement...</div>
+
+    <table v-else-if="users.length > 0" class="product-table">
+      <thead>
+        <tr>
+          <th>Nom d'utilisateur</th>
+          <th>Rôle</th>
+          <th>Magasin</th>
+          <th>Actions</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr v-for="u in users" :key="u.id">
+          <td>{{ u.username }}</td>
+          <td>{{ u.role }}</td>
+          <td>{{ u.magasin }}</td>
+          <td class="td-actions">
+            <button class="btn-edit" @click="openEditModal(u)">Modifier</button>
+            <button class="btn-delete" @click="confirmDelete(u)">Supprimer</button>
+          </td>
+        </tr>
+      </tbody>
+    </table>
+
+    <div v-else class="state-message">Aucun utilisateur trouvé.</div>
+
+    <Modal v-if="showAddUserModal" title="Ajouter un utilisateur" @close="showAddUserModal = false">
+      <AddUserForm
+        @success="handleAddSuccess"
+        @cancel="showAddUserModal = false"
+      />
+    </Modal>
+
+    <Modal v-if="showEditModal" title="Modifier l'utilisateur" @close="showEditModal = false">
+      <EditUserForm
+        :user="selectedUser"
+        @success="handleEditSuccess"
+        @cancel="showEditModal = false"
+      />
+    </Modal>
+  </div>
+</template>
+
+<script setup>
+import { ref } from 'vue'
+import { getUsers, deleteUser } from '@/services/userApi'
+import AddUserForm from '@/components/admin/AddUserForm.vue'
+import EditUserForm from '@/components/admin/EditUserForm.vue'
+import Modal from '@/components/Modal.vue'
+import MessageBox from '@/components/MessageBox.vue'
+
+const users = ref([])
+const loading = ref(false)
+const errorMessage = ref('')
+const successMessage = ref('')
+const showAddUserModal = ref(false)
+const showEditModal = ref(false)
+const selectedUser = ref(null)
+
+async function loadUsers() {
+    console.log('loadUsers appelé')
+  loading.value = true
+  errorMessage.value = ''
+
+  try {
+    const result = await getUsers()
+    users.value = result.data ?? []
+  } catch (err) {
+    errorMessage.value = err.message
+  } finally {
+    loading.value = false
+  }
+}
+
+function openEditModal(u) {
+  selectedUser.value = u
+  showEditModal.value = true
+}
+
+function handleAddSuccess() {
+  showAddUserModal.value = false
+  successMessage.value = 'Utilisateur ajouté avec succès'
+  setTimeout(() => { successMessage.value = '' }, 3000)
+  loadUsers()
+}
+
+function handleEditSuccess(updatedUser) {
+  const index = users.value.findIndex(u => u.id === updatedUser.id)
+  if (index !== -1) {
+    users.value[index] = updatedUser
+  }
+  showEditModal.value = false
+  successMessage.value = 'Utilisateur modifié avec succès'
+  setTimeout(() => { successMessage.value = '' }, 3000)
+}
+
+async function confirmDelete(u) {
+  if (!confirm(`Supprimer "${u.username}" ?`)) return
+
+  try {
+    await deleteUser(u.id)
+    users.value = users.value.filter(u2 => u2.id !== u.id)
+    successMessage.value = 'Utilisateur supprimé avec succès'
+    setTimeout(() => { successMessage.value = '' }, 3000)
+  } catch (err) {
+    errorMessage.value = err.message
+    setTimeout(() => { errorMessage.value = '' }, 3000)
+  }
+}
+
+defineExpose({ loadUsers })
+</script>
+
+<style scoped>
+.content-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 1.5rem;
+}
+
+.content-title {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: var(--color-text-dark);
+}
+
+.btn-add {
+  padding: 0.6rem 1.25rem;
+  background: var(--color-primary);
+  color: white;
+  border: none;
+  border-radius: var(--radius-btn);
+  font-size: 0.9rem;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.btn-add:hover {
+  background: var(--color-primary-dark);
+}
+
+.product-table {
+  width: 100%;
+  border-collapse: collapse;
+  background: white;
+  border-radius: var(--radius-card);
+  overflow: hidden;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.06);
+}
+
+.product-table th {
+  padding: 0.75rem 1rem;
+  text-align: left;
+  font-size: 0.75rem;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.06em;
+  color: var(--color-text-light);
+  border-bottom: 1.5px solid var(--color-border);
+}
+
+.product-table td {
+  padding: 0.85rem 1rem;
+  font-size: 0.9rem;
+  color: var(--color-text-dark);
+  border-bottom: 1px solid var(--color-border);
+}
+
+.product-table tr:last-child td {
+  border-bottom: none;
+}
+
+.td-actions {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.btn-edit {
+  padding: 0.3rem 0.75rem;
+  border: 1.5px solid var(--color-border);
+  background: var(--color-bg-soft);
+  border-radius: 6px;
+  font-size: 0.8rem;
+  cursor: pointer;
+  color: var(--color-text-medium);
+}
+
+.btn-delete {
+  padding: 0.3rem 0.75rem;
+  border: 1.5px solid #fecaca;
+  background: #fff5f5;
+  border-radius: 6px;
+  font-size: 0.8rem;
+  cursor: pointer;
+  color: #dc2626;
+}
+
+.state-message {
+  padding: 2rem;
+  text-align: center;
+  color: var(--color-text-light);
+  background: white;
+  border-radius: var(--radius-card);
+}
+
+.state-message--error {
+  color: #991b1b;
+  background: #fee2e2;
+}
+</style>
